Refactor - useMap 훅 분리
