+++ ePisode/src/pages/MapPage/MapPage.jsx
@@ -1,35 +1,11 @@
-import React, { useEffect } from 'react'
+import React, { useRef } from 'react'
+import useMap from '../../hooks/useMap'
 
 export default function MapPage() {
-  const mapRef = React.useRef(null)
+  const mapRef = useRef(null)
+  const apiKey = import.meta.env.VITE_KAKAO_API_KEY
 
-  useEffect(() => {
-    const mapScript = document.createElement('script')
-
-    mapScript.async = true
-    mapScript.src = `//dapi.kakao.com/v2/maps/sdk.js?appkey=${import.meta.env.VITE_KAKAO_API_KEY}&autoload=false`
-
-    document.head.appendChild(mapScript)
-
-    const onLoadKakaoMap = () => {
-      window.kakao.maps.load(() => {
-        const mapContainer = mapRef.current
-        const mapOption = {
-          center: new window.kakao.maps.LatLng(33.450701, 126.570667),
-          level: 3,
-        }
-        new window.kakao.maps.Map(mapContainer, mapOption)
-      })
-    }
-
-    window.onload = () => {
-      onLoadKakaoMap()
-    }
-
-    return () => {
-      window.onload = null
-    }
-  }, [])
+  useMap(mapRef, apiKey)
 
   return <div ref={mapRef} style={{ width: '100%', height: '100vh' }} id="map"></div>
 }
